+++ tencentcloud/cam/v20190116/errors.go
@@ -26,6 +26,9 @@ const (
 	// 操作访问密钥错误。
 	FAILEDOPERATION_ACCESSKEY = "FailedOperation.Accesskey"
 
+	// 集团策略不允许操作。
+	FAILEDOPERATION_ORGANIZATIONPOLICYOPERATEERROR = "FailedOperation.OrganizationPolicyOperateError"
+
 	// 用户策略数超过上限。
 	FAILEDOPERATION_POLICYFULL = "FailedOperation.PolicyFull"