+++ tencentcloud/essbasic/v20210526/models.go
@@ -5981,13 +5981,17 @@ type FlowResourceUrlInfo struct {
 }
 
 type FormField struct {
-	// 控件填充vaule，ComponentType和传入值类型对应关系：
-	// TEXT - 文本内容
-	// MULTI_LINE_TEXT - 文本内容
-	// CHECK_BOX - true/false
-	// FILL_IMAGE、ATTACHMENT - 附件的FileId，需要通过UploadFiles接口上传获取
-	// SELECTOR - 选项值
-	// DYNAMIC_TABLE - 传入json格式的表格内容，具体见数据结构FlowInfo
+	// 控件填充值，ComponentType和传入值格式对应关系如下：
+	// <ul>
+	// <li>TEXT - 普通文本控件，需输入文本字符串；</li>
+	// <li>MULTI_LINE_TEXT - 多行文本控件，需输入文本字符串；</li>
+	// <li>CHECK_BOX - 勾选框控件，若选中需填写ComponentValue，填写 true或者 false 字符串；</li>
+	// <li>FILL_IMAGE - 图片控件，需填写ComponentValue为图片的资源 ID；</li>
+	// <li>DYNAMIC_TABLE - 动态表格控件；</li>
+	// <li>ATTACHMENT - 附件控件，需填写ComponentValue为附件图片的资源 ID列表，以逗号分割；</li>
+	// <li>DATE - 日期控件；格式为 <b>xxxx年xx月xx日</b> 字符串；</li>
+	// <li>DISTRICT - 省市区行政区控件，需填写ComponentValue为省市区行政区字符串内容；</li>
+	// </ul>
 	ComponentValue *string `json:"ComponentValue,omitnil" name:"ComponentValue"`
 
 	// 表单域或控件的ID，跟ComponentName二选一，不能全为空；